Add localization

+++ src/Oip.WebClient/projects/oip-common/src/components/menu/menu-item.component.ts
@@ -3,22 +3,36 @@ import { NavigationEnd, Router, RouterLinkActive, RouterLink } from '@angular/ro
 import { animate, state, style, transition, trigger } from '@angular/animations';
 import { Subscription } from 'rxjs';
 import { filter } from 'rxjs/operators';
-import { LayoutService } from "./../../services/app.layout.service";
-import { MenuService } from "./../../services/app.menu.service";
+import { LayoutService } from "../../services/app.layout.service";
+import { MenuService } from "../../services/app.menu.service";
 import { RippleModule } from 'primeng/ripple';
 import { NgIf, NgClass, NgFor } from '@angular/common';
-import { MenuItem, MenuItemCommandEvent, PrimeIcons } from "primeng/api";
+import { ConfirmationService, MenuItem, MenuItemCommandEvent, PrimeIcons } from "primeng/api";
 import { MenuItemCreateDialogComponent } from "./menu-item-create-dialog.component";
 import { ContextMenu, ContextMenuModule } from "primeng/contextmenu";
-import { MsgService } from "./../../services/msg.service";
+import { MsgService } from "../../services/msg.service";
 import { MenuItemEditDialogComponent } from "./menu-item-edit-dialog.component";
 import { ContextMenuItemDto } from "../../dtos/context-menu-item.dto";
+import { TranslateService } from "@ngx-translate/core";
+import { ConfirmDialog } from "primeng/confirmdialog";
+
+interface MenuItemComponentTranslation {
+  delete: string;
+  edit: string;
+  new: string;
+  deleteItemConfirmHeader: string;
+  deleteItemConfirmMessage: string;
+  deleteItemSuccessMessage: string;
+  deleteItemConfirmRejectButtonPropsLabel: string;
+  deleteItemConfirmAcceptButtonPropsLabel: string;
+}
 
 @Component({
   // eslint-disable-next-line @angular-eslint/component-selector
   selector: '[app-menuitem]',
   template: `
     <ng-container>
+      <p-confirm-dialog/>
       <div *ngIf="root && item.visible !== false" class="layout-menuitem-root-text"
            (contextmenu)="onContextMenu($event, item)">{{ item.label }}
       </div>
@@ -69,11 +83,15 @@ import { ContextMenuItemDto } from "../../dtos/context-menu-item.dto";
       transition('collapsed <=> expanded', animate('400ms cubic-bezier(0.86, 0, 0.07, 1)'))
     ])
   ],
-  standalone: true,
-  imports: [NgIf, RippleModule, NgClass, RouterLinkActive, RouterLink, NgFor, ContextMenuModule]
+  imports: [NgIf, RippleModule, NgClass, RouterLinkActive, RouterLink, NgFor, ContextMenuModule, ConfirmDialog],
+  providers: [ConfirmationService]
 })
 export class MenuItemComponent implements OnInit, OnDestroy {
-  public layoutService = inject(LayoutService);
+  private readonly layoutService = inject(LayoutService);
+  private readonly translateService = inject(TranslateService);
+  private readonly confirmationService = inject(ConfirmationService)
+  private readonly msgService = inject(MsgService);
+
   @Input() item: ContextMenuItemDto;
   @Input() index!: number;
   @Input() @HostBinding('class.layout-root-menuitem') root!: boolean;
@@ -82,44 +100,37 @@ export class MenuItemComponent implements OnInit, OnDestroy {
   @Input() menuItemEditDialogComponent: MenuItemEditDialogComponent;
   @Input() contextMenu: ContextMenu;
 
-  active = false;
-  menuSourceSubscription: Subscription;
-  menuResetSubscription: Subscription;
+  private active = false;
+  private subscriptions: Subscription[] = [];
+  private localization: MenuItemComponentTranslation = {} as MenuItemComponentTranslation;
 
-  key: string = "";
-  itemMenu: MenuItem[] = [
-    { label: 'New', icon: PrimeIcons.PLUS, command: (event) => this.newClick(event) },
-    { label: 'Edit', icon: PrimeIcons.FILE_EDIT, command: (event) => this.editClick(event) },
-    { separator: true },
-    {
-      label: 'Delete', icon: PrimeIcons.TRASH, command: (event) => {
-        this.deleteItem(event).then();
-      }
-    },
-  ];
-  msgService = inject(MsgService);
+  private key: string = "";
 
   constructor(private readonly cd: ChangeDetectorRef, public router: Router, private readonly menuService: MenuService) {
-    this.menuSourceSubscription = this.menuService.menuSource$.subscribe(value => {
+    this.subscriptions.push(this.menuService.menuSource$.subscribe(value => {
       Promise.resolve(null).then(() => {
         if (value.routeEvent) {
-          this.active = (value.key === this.key || value.key.startsWith(this.key + '-')) ? true : false;
+          this.active = (value.key === this.key || value.key.startsWith(this.key + '-'));
         } else if (value.key !== this.key && !value.key.startsWith(this.key + '-')) {
           this.active = false;
         }
       });
-    });
+    }));
 
-    this.menuResetSubscription = this.menuService.resetSource$.subscribe(() => {
+    this.subscriptions.push(this.menuService.resetSource$.subscribe(() => {
       this.active = false;
-    });
+    }));
 
-    this.router.events.pipe(filter(event => event instanceof NavigationEnd))
+    this.subscriptions.push(this.router.events.pipe(filter(event => event instanceof NavigationEnd))
       .subscribe(params => {
         if (this.item.routerLink) {
           this.updateActiveStateFromRoute();
         }
-      });
+      }));
+
+    this.subscriptions.push(this.translateService.get('menuItemComponent').subscribe((value: MenuItemComponentTranslation) => {
+      this.localization = value;
+    }));
   }
 
   ngOnInit() {
@@ -173,13 +184,7 @@ export class MenuItemComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
-    if (this.menuSourceSubscription) {
-      this.menuSourceSubscription.unsubscribe();
-    }
-
-    if (this.menuResetSubscription) {
-      this.menuResetSubscription.unsubscribe();
-    }
+    this.subscriptions.map(s => s.unsubscribe());
   }
 
   private newClick(e: MenuItemCommandEvent) {
@@ -188,14 +193,47 @@ export class MenuItemComponent implements OnInit, OnDestroy {
 
   onContextMenu($event: MouseEvent, item: any) {
     this.menuService.contextMenuItem = item;
-    this.contextMenu.model = this.itemMenu;
+    this.contextMenu.model = [
+      {
+        label: this.localization.new,
+        icon: PrimeIcons.PLUS,
+        command: (event) => this.newClick(event)
+      },
+      {
+        label: this.localization.edit,
+        icon: PrimeIcons.FILE_EDIT,
+        command: (event) => this.editClick(event)
+      },
+      { separator: true },
+      {
+        label: this.localization.delete,
+        icon: PrimeIcons.TRASH,
+        command: (event) => this.deleteItem(event)
+      },
+    ];
     this.contextMenu.show($event);
   }
 
-  private async deleteItem(event: MenuItemCommandEvent) {
-    await this.menuService.deleteItem(this.menuService.contextMenuItem?.moduleInstanceId);
-    this.msgService.success('Saved security');
-    await this.menuService.loadMenu();
+  private deleteItem(event: MenuItemCommandEvent) {
+    this.confirmationService.confirm({
+      header: this.localization.deleteItemConfirmHeader,
+      message: this.localization.deleteItemConfirmMessage,
+      icon: PrimeIcons.TRASH,
+      rejectButtonProps: {
+        label: this.localization.deleteItemConfirmRejectButtonPropsLabel,
+        severity: 'secondary',
+        outlined: true,
+      },
+      acceptButtonProps: {
+        label: this.localization.deleteItemConfirmAcceptButtonPropsLabel,
+        severity: 'danger',
+      },
+      accept: async () => {
+        await this.menuService.deleteItem(this.menuService.contextMenuItem?.moduleInstanceId);
+        this.msgService.success(this.localization.deleteItemSuccessMessage);
+        await this.menuService.loadMenu();
+      },
+    });
   }
 
   private editClick(event: MenuItemCommandEvent) {
